requirements.txt
langchain
langchain_core
langchain_community
streamlit==1.24.0
streamlit_extras
python-dotenv
fastapi
uvicorn
langchain_huggingface
sse_starlette
bs4
pypdf
chromadb
faiss-cpu
arxiv
groq
langchain-groq
pdfkit
WeasyPrint
reportlab
fpdf
sentence-transformers
PyMuPDF
pytesseract 
pillow 
opencv-python
pdfplumber
spacy
duckduckgo-search